serving/faiss_cpu/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
numpy==1.26.4
faiss-cpu==1.8.0
